# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DHL - Dale Hollow Lake API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pins:
    get:
      operationId: getPins
      tags: [pins]
      summary: Get all pins on the lake
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [fishing_spot, cliff, waterfall, rope_swing, shallow_water, tubing, skiing, houseboat, divers, landmark, hazard, marina, campsite, other]
        - name: profileUserId
          in: query
          required: false
          description: When set, returns the given user's pins for display on their profile (includes their friends-only pins).
          schema:
            type: integer
        - name: lakeId
          in: query
          required: false
          description: Only include pins on this lake
          schema:
            type: integer
      responses:
        "200":
          description: List of pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pin"

    post:
      operationId: createPin
      tags: [pins]
      summary: Drop a pin on the lake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "201":
          description: Pin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
  /pins/{pinId}:
    get:
      operationId: getPin
      tags: [pins]
      summary: Get a specific pin
      parameters:
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"

    delete:
      operationId: deletePin
      tags: [pins]
      summary: Delete a pin
      parameters:
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /pins/{pinId}/like:
    post:
      operationId: likePin
      tags: [pins]
      summary: Like or upvote a pin
      parameters:
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
...
  /pins/{pinId}/favorite:
    post:
      operationId: toggleFavoritePin
      tags: [pins]
      summary: Toggle favorite on a pin
      parameters:
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated pin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
  /pins/pending/approval:
    get:
      operationId: getPendingPins
      tags: [pins]
      summary: Get community pins awaiting owner approval
      responses:
        "200":
          description: List of pending community pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pin"
        "403":
          description: Not the app owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /pins/{pinId}/approve:
    post:
      operationId: approvePin
      tags: [pins]
      summary: Approve a community pin (app owner only)
      parameters:
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "403":
          description: Not the app owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /dock-labels:
    get:
      operationId: getDockLabels
      tags: [dockLabels]
      summary: Get all dock labels on the lake
      parameters:
        - name: lakeId
          in: query
          required: false
          description: Only include dock labels on this lake
          schema:
            type: integer
      responses:
        "200":
          description: List of dock labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DockLabel"

    post:
      operationId: createDockLabel
      tags: [dockLabels]
      summary: Place a dock label (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DockLabelInput"
      responses:
        "201":
          description: Dock label created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DockLabel"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /dock-labels/{labelId}:
    delete:
      operationId: deleteDockLabel
      tags: [dockLabels]
      summary: Remove a dock label (admin only)
      parameters:
        - name: labelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /pins/hazards/active:
    get:
      operationId: getActiveHazards
      tags: [pins]
      summary: Get active (non-expired) hazard pins
      parameters:
        - name: lakeId
          in: query
          required: false
          description: Only include hazards on this lake
          schema:
            type: integer
      responses:
        "200":
          description: List of active hazard pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pin"
...
components:
  schemas:
...
    Badge:
      type: object
      required: [key, label, description, earned]
      properties:
        key:
          type: string
        label:
          type: string
        description:
          type: string
        earned:
          type: boolean
    Rank:
      type: object
      required: [key, title, tier, earnedCount, totalCount]
      properties:
        key:
          type: string
        title:
          type: string
        tier:
          type: integer
        earnedCount:
          type: integer
        totalCount:
          type: integer
        nextTitle:
          type: ["string", "null"]
        nextNeeded:
          type: ["integer", "null"]
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        hometown:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        relationshipStatus:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        work:
          type: ["string", "null"]
        isOnline:
          type: boolean
        isBusiness:
          type: boolean
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        lastSeen:
          type: ["string", "null"]
        lakeStatus:
          type: ["string", "null"]
        boatName:
          type: ["string", "null"]
        boatColor:
          type: ["string", "null"]
        boatType:
          type: ["string", "null"]
        boatBrand:
          type: ["string", "null"]
        boatModel:
          type: ["string", "null"]
        boatYear:
          type: ["integer", "null"]
        boatPhotoUrl:
          type: ["string", "null"]
        homeMarina:
          type: ["string", "null"]
        showBoat:
          type: boolean
        boatNeon:
          type: ["boolean", "null"]
        boatFlag:
          type: ["boolean", "null"]
        boatAccent:
          type: ["string", "null"]
        fleet:
          type: array
          description: The user's boats/watercraft (redacted for other viewers when showBoat=false)
          items:
            $ref: "#/components/schemas/Boat"
        interests:
          type: array
          items:
            type: string
        favoriteThings:
          type: array
          description: Pinned favorites shown on the profile (label + value pairs)
          items:
            $ref: "#/components/schemas/FavoriteThing"
        shareLocation:
          type: boolean
        locationSharingExpiresAt:
          type: ["string", "null"]
        isSharingLocation:
          type: boolean
        requireFollowApproval:
          type: boolean
        showFollowers:
          type: boolean
        showFriends:
          type: boolean
        followerSeeLocation:
          type: boolean
        followerSeePosts:
          type: boolean
        followerSendMessages:
          type: boolean
        showMatureContent:
          type: boolean
        isAdmin:
          type: boolean
        demoMode:
          type: boolean
        isSuspended:
          type: boolean
        warningCount:
          type: integer
        waiverAcceptedAt:
          type: ["string", "null"]
        waiverVersion:
          type: ["string", "null"]
        termsAcceptedAt:
          type: ["string", "null"]
        termsVersion:
          type: ["string", "null"]
        friendStatus:
          type: string
          enum: [none, self, accepted, pending_out, pending_in, blocked, blocked_by]
        followerCount:
          type: integer
        followingCount:
          type: integer
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        rank:
          $ref: "#/components/schemas/Rank"
        primaryLakeId:
          type: integer
          description: The user's home lake (from the static lakes catalog)
        currentLakeId:
          type: ["integer", "null"]
          description: The lake of the user's most recent check-in
        createdAt:
          type: string
...
    FavoriteThing:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
...
    Pin:
      type: object
      required: [id, userId, lat, lng, type, title, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        lakeId:
          type: integer
        lat:
          type: number
        lng:
          type: number
        type:
          type: string
          enum: [fishing_spot, cliff, waterfall, rope_swing, shallow_water, tubing, skiing, houseboat, divers, landmark, hazard, marina, campsite, other]
        title:
          type: string
        description:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [friends, public, community]
        imageUrl:
          type: ["string", "null"]
        approved:
          type: boolean
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        expiresAt:
          type: ["string", "null"]
        likeCount:
          type: integer
        likedByMe:
          type: boolean
        favoritedByMe:
          type: boolean
        isMature:
          type: boolean
        createdAt:
          type: string
    PinInput:
      type: object
      required: [lat, lng, type, title]
      properties:
        lat:
          type: number
        lng:
          type: number
        type:
          type: string
          enum: [fishing_spot, cliff, waterfall, rope_swing, shallow_water, tubing, skiing, houseboat, divers, landmark, hazard, marina, campsite, other]
        title:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [friends, public, community]
        imageUrl:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        severity:
          type: string
          enum: [low, medium, high]
        expiresAt:
          type: ["string", "null"]
        lakeId:
          type: ["integer", "null"]
          description: Which lake the pin belongs to (defaults to Dale Hollow Lake)
    DockLabel:
      type: object
      required: [id, userId, label, lat, lng, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        lakeId:
          type: integer
        label:
          type: string
        emoji:
          type: string
          nullable: true
        lat:
          type: number
        lng:
          type: number
        createdAt:
          type: string
...
    DockLabelInput:
      type: object
      required: [label, lat, lng]
      properties:
        label:
          type: string
        emoji:
          type: string
        lat:
          type: number
        lng:
          type: number
        lakeId:
          type: ["integer", "null"]
          description: Which lake the dock label belongs to (defaults to Dale Hollow Lake)
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Boat:
      type: object
      required: [id, userId, name, boatType, color, isPrimary, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        boatType:
          type: string
        color:
          type: string
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
        neon:
          type: boolean
        flag:
          type: boolean
        accent:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        horsepower:
          type: ["integer", "null"]
        engineInfo:
          type: ["string", "null"]
        lengthFt:
          type: ["integer", "null"]
        favoriteMarina:
          type: ["string", "null"]
        favoriteCove:
          type: ["string", "null"]
        favoriteActivity:
          type: ["string", "null"]
        mods:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        createdAt:
          type: string
...