# lib/api-zod/src/generated/types/dockLabelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DHL - Dale Hollow Lake API
 * OpenAPI spec version: 0.1.0
 */

export interface DockLabelInput {
  label: string;
  emoji?: string;
  lat: number;
  lng: number;
  /**
     * Which lake the dock label belongs to (defaults to Dale Hollow Lake)
     * @nullable
     */
  lakeId?: number | null;
}