# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DHL - Dale Hollow Lake API
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from './badge';
import type { Boat } from './boat';
import type { FavoriteThing } from './favoriteThing';
import type { Rank } from './rank';
import type { UserFriendStatus } from './userFriendStatus';

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  hometown?: string | null;
  /** @nullable */
  birthday?: string | null;
  /** @nullable */
  relationshipStatus?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  work?: string | null;
  isOnline?: boolean;
  isBusiness?: boolean;
  /** @nullable */
  currentLat?: number | null;
  /** @nullable */
  currentLng?: number | null;
  /** @nullable */
  lastSeen?: string | null;
  /** @nullable */
  lakeStatus?: string | null;
  /** @nullable */
  boatName?: string | null;
  /** @nullable */
  boatColor?: string | null;
  /** @nullable */
  boatType?: string | null;
  /** @nullable */
  boatBrand?: string | null;
  /** @nullable */
  boatModel?: string | null;
  /** @nullable */
  boatYear?: number | null;
  /** @nullable */
  boatPhotoUrl?: string | null;
  /** @nullable */
  homeMarina?: string | null;
  showBoat?: boolean;
  /** @nullable */
  boatNeon?: boolean | null;
  /** @nullable */
  boatFlag?: boolean | null;
  /** @nullable */
  boatAccent?: string | null;
  /** The user's boats/watercraft (redacted for other viewers when showBoat=false) */
  fleet?: Boat[];
  interests?: string[];
  /** Pinned favorites shown on the profile (label + value pairs) */
  favoriteThings?: FavoriteThing[];
  shareLocation?: boolean;
  /** @nullable */
  locationSharingExpiresAt?: string | null;
  isSharingLocation?: boolean;
  requireFollowApproval?: boolean;
  showFollowers?: boolean;
  showFriends?: boolean;
  followerSeeLocation?: boolean;
  followerSeePosts?: boolean;
  followerSendMessages?: boolean;
  showMatureContent?: boolean;
  isAdmin?: boolean;
  demoMode?: boolean;
  isSuspended?: boolean;
  warningCount?: number;
  /** @nullable */
  waiverAcceptedAt?: string | null;
  /** @nullable */
  waiverVersion?: string | null;
  /** @nullable */
  termsAcceptedAt?: string | null;
  /** @nullable */
  termsVersion?: string | null;
  friendStatus?: UserFriendStatus;
  followerCount?: number;
  followingCount?: number;
  badges?: Badge[];
  rank?: Rank;
  /** The user's home lake (from the static lakes catalog) */
  primaryLakeId?: number;
  /**
     * The lake of the user's most recent check-in
     * @nullable
     */
  currentLakeId?: number | null;
  createdAt: string;
}